Updated AI / Search
Added the A* algorithm to the game, which allows the player to search
for the exit / goal node in the maze from any point in the maze.
A few bugs still exist with the implementation...
Added new images to be used later
Added fuzzy logic files to be implemented later
Unit tests will be implemented later
That's all for now!