Added Docs, Updated the AI & Small Fixes
Added the doc files and updated the enemy boss AI to hunt and kill the
player when within 60 steps of the goal node in the maze.
Now composing the main enemy and player placement code in their own
class.
Cleaned up some of the code & small bug fixes
CPU usage is quite high at times, could be related to the amount of
enemies spawned in-game. Or could be the A* implementation in boss
enemies.